emacs: few improvements
- set indentation for ledger files
- enable the undo module for doom

Unfortunately, there seems to be a lack of a uniform way to configure
indentation in emacs. Each major mode has its own way of configuring
indentation, and many of them seem to confuse 'tab-width' [1] with
indentation. This makes it nearly impossible to have separate settings
for 'tab-width', 'tab-stop-list' [2], and 'standard-indent' [3] without
breaking things. So I'll just stick with having them all set to '2' for
now.

[1]: Roughly equal to 'tabstop' in Vim
[2]: Roughly equal to 'softtabstop' in Vim
[3]: Roughly equal to 'shiftwidth' in Vim