patch(tmux): new patch for 3.1, and a few more improvements
tmux version 3.1 brings alongs a number of goodies including:

- regex search in copy-mode (no more need for copycat!!)
- the ability to add an description to keybindings
- ~/.config/tmux/tmux.conf as an alternative path for the config
- and quite a few more that I haven't yet checked

This commit replaces most usages of copycat with the builtin regex
search. More updates for 3.1 would come later. Relocation of tmux.conf
would be put on hold until the Ubuntu package reaches 3.1.

I've also improved the patches so that it would no longer be necessary to
apply multiple patches to obtain a working configuration for a given
version. From now on, to obtain a configuration for 3.0, you'd only need
to apply the patch for 3.0, and not for 2.8. The "pre-" and "post-"
prefixes are removed from the patch filenames to reflect this change.