fork.sh: fix portability issues
Follow up of ad213dc. fork.sh is one of the very few scripts that
requires a higher degree of portability without relying on git branches,
since it's only intended to be run from the master branch. With that
being said, uses of sed were causing portability issues because of
BSD/GNU incompabilites which arise when you target both Linux and macOS.
Therefore, replace sed with perl to fix this problem.