zsh: support prompt themes in the portable config
Add support for prompt themes "concise" and "dashboard." They can both
be enabled with the 'prompt' command. For example, to enable the
"dashboard" theme, run the following command:

    prompt dashboard

To keep things on this branch as small as possible, the actual code for
the prompts won't be included in the dotfiles archive itself.  They'll
be downloaded for each zsh session when you actually attempt to use
them.

The base URL for downloading additional files like these can configured
with zstyle:

    zstyle ':dotfiles' baseurl 'https://dotfiles.example/dotfiles'