setup: install coc-ultisnips
coc-ultisnips is needed because coc-snippets fails to process many
UltiSnips snippets from honza/vim-snippets. Removing coc-snippets isn't
an option either, because we'd then lose LSP snippets and snippets
preview. The keybindings remain the same, <Tab> for coc-snippets and
<CTRL-X><CTRL-J> for ultisnips.

The installation would be done only when the '--init' flag is specified,
since it's unfortunately not possible to manage it using plugin
managers.