shell: add .hushlogin to prevent potential slowdown of shell startup
When launching a new shell in macOS, the /usr/bin/login program goes
through the system log files to find the last login time in order to
display the "Last Login:" message. Depending on the size of the system
logs [1], this process may significantly slow down the shell startup
time.

To suppress this message, users can either:

1. create an empty ~/.hushlogin file
2. pass the -q flag to /usr/bin/login when starting a new shell
3. specify the shell manually

Unfornately, all of these approaches has problems:

|   | problems                                                  |
|---+-----------------------------------------------------------|
| 1 | this won't stop the log lookup process                    |
| 2 | it's not possible for users to specify the flags directly |
| 3 | it's rather inconvinient                                  |

Because of this, I've ignored this problem up until now. However, I
recently found out that iTerm2 has a nice solution for this problem
[2][3]. In short, it takes approach 2 when ~/.hushlogin is detected.
Since I use iTerm2 all the time, I decided to place the .hushlogin file
to take advantage of this feature.

[1]: https://apple.stackexchange.com/q/41743
[2]: https://gitlab.com/gnachman/iterm2/-/issues/4425
[3]: https://github.com/gnachman/iTerm2/commit/f3308eb14c53dd63f1d2bab376fdd311a3a0bcfe