git: improve diff output with smarter hunk header generation
Use pre-defined filetype-specific regexes for hunk header generation.
Since the filetype is determined solely by the file extension, it won't
have terrifc accuracy. However, it should be good enough as a fallback
for repositories that doesn't include its own settings in the
.gitattributes file.

More detailed description including the list of supported filetypes can
be found in the official docs:
https://git-scm.com/docs/gitattributes#_defining_a_custom_hunk_header