home-manager: locale fixes for glibc
- Add a new option, 'dotfiles.fixes.localeArchive.enable'

When enabled, this would attempt to fix incompabilities arising from the
use of the system-provided locale archive.

- Add a new optional wrapper package for zsh

This fixes the locale problem mentioned above. Intended for use with
stable Linux distributions with an outdated zsh package and should be
usable as a login shell.

Since the majority of popular distributions have reasonably recent
versions of zsh, no option is provided to install this package. To
install, include the following in ~/.config/nixpkgs/home.d/config.nix:

    home.packages = [ (import ./dotfiles/pkgs { }).zsh ];