chore(emacs): adjust titlebar settings for emacs >= 27.1
The 'frame-title-format' option in Emacs < 27.1 used to show an icon of
the current file in the titlebar on macOS when it was set to 't'. Now,
this behavior is controlled by a different option, 'ns-use-proxy-icon'
and the 'frame-title-format' behaves in line with other platforms [1].

Following the above, this change reverts the macOS specific settings
made to the 'frame-title-format'. Also, it doesn't set
'ns-use-proxy-icon' because it seems to be enabled by default.

[1]: https://www.gnu.org/savannah-checkouts/gnu/emacs/news/NEWS.27.1