Revert "docker: stop docker from swallowing ctrl-p"
This reverts commit 4256c5e1d1ffdc3300b74b5c7d58fca81e77f776.

The configuration was there to prevent Docker from hijacking the
widely-used CTRL-P keybinding and reduce frustrations with interactive
use. However, it proved to be less than useful for a number of reasons:

- Docker places runtime state in 'config.json', making it painful to
  manage with Git
- 'config.json' would have to be placed under /root to have any effect
  in a properly configured Linux environment where the Docker CLI needs
  to be invoked as root

The alternative is to add the '--detach-keys "ctrl-q,d"' flag from the
command line. While unfortunate, this may not be so bad after all
because it's frequently necessary to pass a long list of command line
flags to the Docker CLI even without '--detach-keys'.