fix(emacs): install emacsql via Nix
Recent versions of emacsql attempts to build emacsql-sqlite.c at
runtime. I'm not sure why, considering that its companion package,
emacsql-sqlite, builds the same code too.

But whatever the reason, it has to be moved over to Nix. Otherwise, an
entire C toolchain would have to be exposed to the user environment in
order to create an impure binary.